change task dao to use long

+++ Smart Task Tracker/src/main/java/com/mthree/dao/TaskDao.java
@@ -1,14 +1,15 @@
 package com.mthree.dao;
 
-import com.mthree.entity.Category;
-import com.mthree.entity.Task;
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
+import com.mthree.entity.Category;
+import com.mthree.entity.Task;
 
 @Repository
-public interface TaskDao extends JpaRepository<Task, Integer> {
+public interface TaskDao extends JpaRepository<Task, Long> {
 
     /*
     *List tasks given category
